+++ kDrive/Utils/AppFactoryService.swift
@@ -20,107 +20,29 @@ import Foundation
 import InfomaniakDI
 import kDriveCore
 import os.log
-import UIKit
 
 /// Something that loads the DI on init
 public struct EarlyDIHook {
     public init(context: DriveAppContext) {
         os_log("EarlyDIHook")
 
-        let extraDependencies = [
+        var extraDependencies = [
             Factory(type: NavigationManageable.self) { _, _ in
                 NavigationManager()
             },
             Factory(type: AppContextServiceable.self) { _, _ in
                 AppContextService(context: context)
-            },
-            Factory(type: AppRestorationService.self) { _, _ in
-                AppRestorationService()
             }
         ]
+
+        #if !ISEXTENSION
+        // AppRestorationService only available for main app, not extensions
+        extraDependencies += [Factory(type: AppRestorationService.self) { _, _ in
+            AppRestorationService()
+        }]
+        #endif
 
         // setup DI ASAP
         FactoryService.setupDependencyInjection(other: extraDependencies)
     }
 }
-
-// TODO: Refactor with Scenes / NSUserActivity
-#if !ISEXTENSION
-
-public final class AppRestorationService {
-    /// Path where the state restoration state is saved
-    private static let statePath = FileManager.default
-        .urls(for: .libraryDirectory, in: .userDomainMask)
-        .first?
-        .appendingPathComponent("Saved Application State")
-
-    @LazyInjectService private var accountManager: AccountManageable
-
-    /// State restoration version
-    private static let currentStateVersion = 4
-
-    /// State restoration key
-    private static let appStateVersionKey = "appStateVersionKey"
-
-    public init() {
-        // META: keep SonarCloud happy
-    }
-
-    public func shouldSaveApplicationState(coder: NSCoder) -> Bool {
-        Log.appDelegate("shouldSaveApplicationState")
-        Log.appDelegate("Restoration files:\(String(describing: Self.statePath))")
-        coder.encode(Self.currentStateVersion, forKey: Self.appStateVersionKey)
-        return true
-    }
-
-    public func shouldRestoreApplicationState(coder: NSCoder) -> Bool {
-        let encodedVersion = coder.decodeInteger(forKey: Self.appStateVersionKey)
-        let shouldRestoreApplicationState = Self.currentStateVersion == encodedVersion &&
-            !(UserDefaults.shared.legacyIsFirstLaunch || accountManager.accounts.isEmpty)
-        Log.appDelegate("shouldRestoreApplicationState:\(shouldRestoreApplicationState)")
-        return shouldRestoreApplicationState
-    }
-
-    public func reloadAppUI(for drive: Drive) {
-        accountManager.setCurrentDriveForCurrentAccount(drive: drive)
-        accountManager.saveAccounts()
-
-        guard let currentDriveFileManager = accountManager.currentDriveFileManager else {
-            return
-        }
-
-        // Read the last tab selected in order to properly reload the App's UI.
-        // This should be migrated to NSUserActivity at some point
-        let lastSelectedTab = UserDefaults.shared.lastSelectedTab
-        let newMainTabViewController = MainTabViewController(
-            driveFileManager: currentDriveFileManager,
-            selectedIndex: lastSelectedTab
-        )
-        (UIApplication.shared.delegate as? AppDelegate)?.setRootViewController(newMainTabViewController)
-
-        // cleanup
-        UserDefaults.shared.lastSelectedTab = nil
-    }
-}
-
-#else
-
-public final class AppRestorationService {
-    public init() {
-        // META: keep SonarCloud happy
-    }
-
-    public func shouldSaveApplicationState(coder: NSCoder) -> Bool {
-        false
-    }
-
-    public func shouldRestoreApplicationState(coder: NSCoder) -> Bool {
-        false
-    }
-
-    public func reloadAppUI(for drive: Drive) {
-        // NOOP
-    }
-}
-
-#endif
